+++ driver/web/apis_admin.go
@@ -146,22 +146,22 @@ func (h AdminApisHandler) GetMessages(l *logs.Log, r *http.Request, claims *toke
 // @Security AdminUserAuth
 // @Router /admin/message [post]
 func (h AdminApisHandler) CreateMessage(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HTTPResponse {
-	var inputMessage Def.SharedReqCreateMessage
-	err := json.NewDecoder(r.Body).Decode(&inputMessage)
+	var inputData Def.SharedReqCreateMessage
+	err := json.NewDecoder(r.Body).Decode(&inputData)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionDecode, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, true)
 	}
 
 	orgID := claims.OrgID
 	appID := claims.AppID
-
-	time, priority, subject, body, inputData, inputRecipients, recipientsCriteria, recipientsAccountCriteria, topic := getMessageData(inputMessage)
-
 	sender := model.Sender{Type: "user", User: &model.CoreAccountRef{UserID: claims.Subject, Name: claims.Name}}
 
-	message, err := h.app.Services.CreateMessage(orgID, appID,
-		sender, time, priority, subject, body, inputData, inputRecipients, recipientsCriteria,
-		recipientsAccountCriteria, topic, false)
+	inputMessage := getMessageData(inputData)
+	inputMessage.OrgID = orgID
+	inputMessage.AppID = appID
+	inputMessage.Sender = sender
+
+	message, err := h.app.Services.CreateMessage(inputMessage)
 	if err != nil {
 		return l.HTTPResponseErrorAction(logutils.ActionCreate, "message", nil, err, http.StatusInternalServerError, true)
 	}