+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.14.0] - 2023-03-31
+### Added
+- Add and delete recepient to message [#147](https://github.com/rokwire/notifications-building-block/issues/147)
+- Expose create and delete many messages BBs APIs [#148](https://github.com/rokwire/notifications-building-block/issues/148)
+
 ## [1.13.0] - 2023-02-21
 ### Added
 - Expose cancel message building block API [#144](https://github.com/rokwire/notifications-building-block/issues/144)